driver/misc: aspeed-espi: Reserve memory cycle for MMBI
The Aspeed MMBI device conducts data transfer over
the memory cycle of eSPI peripheral channel. Hence
when MMBI is enabled, the memory cycle function is
reserved for MMBI use.

This patch allows the eSPI peripheral driver detects
the activation of the MMBI device and thus leave the
memory cycle untouched.