soc: aspeed: Add mailbox support
Mailbox can used for the communication between the Host and the BMC.
The interrupts of signaling each other are also supported. There are
total 16 data registers implemented for AST2400/AST2500. For AST2600,
the number of the data registers are extended to 32.

This patch add a general driver for the support of the Aspeed mailbox
controller.

Signed-off-by: Chia-Wei, Wang <chiawei_wang@aspeedtech.com>