i3c: setup device role according to the property "secondary"
the i3c subsystem does not yet define the property for configuration of
a secondary master.  Temporarily, add a bool property "secondary" to
indicate if the current device is going to be a secondary master (true)
or a main master (false).

The register DEVICE_CTRL_EXTENDED (offset 0xb0) bit[1:0] configures the
device role. When the device is going to be initialized as a secondary
master, bit[1:0] shall be set to 2b'01 (slave)

Signed-off-by: Dylan Hung <dylan_hung@aspeedtech.com>
Change-Id: I00a55e8222af2187598fadaa22a5b4b38835c2c4