i3c: create device data space for the secondary master
Generally, the divice data space is created in the common code
i3c_master_attach_i3c_dev. But that function does not create the space
for the master device itself.  For secondary master, it needs for the
device data space to store the slave message info, so we manually create
the space here.

Signed-off-by: Dylan Hung <dylan_hung@aspeedtech.com>
Change-Id: I24e7223020250a7520798243c80c027506a386ac