i3c: fix SETDASA fail for standard MIPI device.
When a slave is declared in the DT with static address and assigned
dynamic address, the subsystem would reserved the address slot for its
dynamic address first.  If the static address equals to the dynamic
address, an error would occur in i3c_master_attach_i3c_dev since the
address slot for the static address has already been used. Skip
reserving the address slot if these address are the same.

Signed-off-by: Dylan Hung <dylan_hung@aspeedtech.com>
Change-Id: Ic98e5792eb6b3b7f53ccf36f1ad7aee0ce924b14