media: aspeed: Support aspeed mode to reduce compressed data
aspeed support differential jpeg format which only compress the parts
which are changed. In this way, it reduces both the amount of data to be
transferred by network and those to be decoded on the client side.

4 new ctrls are added:
*Aspeed JPEG Format: to control aspeed's partial jpeg on/off
*Aspeed Compression Mode: to control aspeed's compression mode
*Aspeed HQ Mode: to control aspeed's HQ mode on/off
*Aspeed HQ Quality: to control the quality of aspeed's HQ mode

Aspeed JPEG Format requires an additional buffer, called bcd, to store
the information that which macro block in the new frame is different
from the old one.

To have bcd correctly working, we need to swap the buffers for src0/1 to
make src1 refer to previous frame and src0 to the coming new frame.

Change-Id: I034ab9c2fa95413fd1fd7e6a878969fff0f94f21
Signed-off-by: Jammy Huang <jammy_huang@aspeedtech.com>