soc: aspeed: mctp: Fix cpu stall issue.
Add the condition to break the while loop which is used to find the
valid packet from RX pool. When the pcie interrupt comes in during the
program is looking for a valid packet, all of the packet will be marked
as invalid and CPU will stuck in the loop.

Signed-off-by: Billy Tsai <billy_tsai@aspeedtech.com>
Change-Id: I61e17e65110b1688e94f9dfda835152bb2a40ddf