i3c: aspeed: revise PP timing for JESD403 bus context
The original design for multiple SCL supporting is done by setting
EXT_LCNT_TIMING register, and select the corresponding speed index in
the transfer command.  The high period of the SCL is fixed for all SCL
rates. However, when the long trace or heavy loading on the SCL pin,
the hold time of the SCL high period may not long enough to perform a
square wave even if we slow down the SCL.  This patch sets up the
typical SCL PP rate according to i3c-scl-hz property, and configure the
PP timing by the predefined lookup table.  It makes the master
controller uses the same SCL frequency for all the slave devices with
long enough high period.

Signed-off-by: Dylan Hung <dylan_hung@aspeedtech.com>
Change-Id: I5e2ca5fa13a0f5cf63b0222d67a90cd3e8244638